Ensure TypingIndicator wraps correctly with very long names (#1123)
