+++ change/@internal-acs-ui-common-671256ec-9252-4f5f-9530-f9e70e608b70.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Move function outside of this packlet",
+  "packageName": "@internal/acs-ui-common",
+  "email": "2684369+JamesBurnside@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
