+++ change/@internal-react-components-5d12798f-7dd1-4124-85ed-f97f1b114159.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "Ensure typing indicator wraps correctly with long names",
+  "packageName": "@internal/react-components",
+  "email": "2684369+JamesBurnside@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
