+++ change/@internal-react-composites-4e817ead-53d0-426d-8c03-0ffde5989f9b.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Update snapshots",
+  "packageName": "@internal/react-composites",
+  "email": "2684369+JamesBurnside@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
