+++ change/@internal-storybook-5bfcb168-ef80-4f24-9b45-b19b0071faf7.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Fixup story",
+  "packageName": "@internal/storybook",
+  "email": "2684369+JamesBurnside@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
